Database transaction lock & handle deadlock

+++ db/migration/000001_init_schema.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE "accounts" (
                             "id" bigserial PRIMARY KEY,
                             "owner" varchar NOT NULL,
-                            "balance" int NOT NULL,
+                            "balance" bigint NOT NULL,
                             "currency" varchar NOT NULL,
                             "created_at" timestamp NOT NULL DEFAULT (now())
 );
